Ajout de commentaires, suppression de fichiers inutiles

+++ CoreEngine/Window.java
@@ -4,6 +4,10 @@ import javax.swing.JFrame;
 
 import GraphicsEngine.GraphicsSystem;
 
+/*
+ * Singleton responsable de la gestion de la fenêtre
+ * Initialisation de la JFrame et lien avec les Manager autres (Keyboard, Mouse, GraphicsSystem)
+ */
 public class Window extends JFrame {
     private Window() {
         setTitle("C&G Engine");
@@ -23,11 +27,11 @@ public class Window extends JFrame {
         return instance;
     }
 
-    public float GetWidth() {
+    public int GetWidth() {
         return width;
     }
 
-    public float GetHeight() {
+    public int GetHeight() {
         return height;
     }
